Consolidate time util
